refactor: compose domain models into route directory
