+++ api/domain/src/model.rs
@@ -1,14 +1,10 @@
-pub use coordinate::Coordinate;
-pub use operation::{Operation, OperationType};
 pub use route::{Route, RouteInfo};
-pub use segment::{Segment, SegmentList};
+// TODO: この辺をmodel/route.rsに移動して、外からもmodel::route::でアクセスさせるようにする
+pub use route::coordinate::Coordinate;
+pub use route::operation::{Operation, OperationType};
+pub use route::route_gpx::RouteGpx;
+pub use route::segment_list::{Segment, SegmentList};
 pub use types::{Distance, Elevation, Latitude, Longitude, Polyline, RouteId};
 
-pub use self::gpx::RouteGpx;
-
-pub(crate) mod coordinate;
-pub(crate) mod gpx;
-pub(crate) mod operation;
 pub(crate) mod route;
-pub(crate) mod segment;
 pub(crate) mod types;
